Minor JS cleanup and add compiled CSS
This noob learned that SCSS is server-side script and the browser is served the compiled CSS, so I brought the compiled CSS over from CodePen. I still need to clean it up so that it displays properly -- there are some issues, mainly with spacing, especially on the control pad.

Also read through the JavaScript to remind myself how the game works, and made a few very minor edits. Mainly took a few instances of loose equality checks and switched them to strict equality.